refactor: merge `pointcloud_map_loader_node_core` into `pointcloud_map_loader`
  * For file name rule consistency
