+++ map/autoware_map_loader/CMakeLists.txt
@@ -16,7 +16,6 @@ ament_auto_add_library(pointcloud_map_loader_node SHARED
   src/pointcloud_map_loader/differential_map_loader_module.cpp
   src/pointcloud_map_loader/selected_map_loader.cpp
   src/pointcloud_map_loader/selected_map_loader_module.cpp
-  src/pointcloud_map_loader/pointcloud_map_loader_node_core.cpp
   src/pointcloud_map_loader/utils.cpp
 )
 target_link_libraries(pointcloud_map_loader_node ${PCL_LIBRARIES})
